Extended eCommerce Service to allow for additional mutations.

+++ ecommerce_service/src/main/java/at/technikum/masterproject/ecommerceservice/purchase/PurchaseItemRepository.java
@@ -3,6 +3,6 @@ package at.technikum.masterproject.ecommerceservice.purchase;
 import at.technikum.masterproject.ecommerceservice.purchase.model.PurchaseItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseItemRepository extends JpaRepository<PurchaseItem, Long> {
+public interface PurchaseItemRepository extends JpaRepository<PurchaseItem, Integer> {
 
 }
