+++ ecommerce_service/src/main/java/at/technikum/masterproject/ecommerceservice/purchase/PurchaseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface PurchaseRepository extends JpaRepository<Purchase, Long> {
+interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
 
   List<Purchase> findPurchasesByCustomerId(Integer customerId);
 }
